@@ -18,8 +18,8 @@ resolver = "2"
 [workspace.package]
 authors = ["Robin Freyler <robin.freyler@gmail.com>"]
 repository = "https://github.com/wasmi-labs/wasmi"
-version = "0.46.0"
-rust-version = "1.83"
+version = "0.47.0"
+rust-version = "1.86"
 edition = "2021"
 readme = "README.md"
 license = "MIT/Apache-2.0"
@@ -30,15 +30,15 @@ exclude = ["tests"]
 [workspace.dependencies]
 
 # Wasmi crates
-wasmi = { version = "0.46.0", path = "crates/wasmi", default-features = false }
-wasmi_wasi = { version = "0.46.0", path = "crates/wasi", default-features = false }
-wasmi_core = { version = "0.46.0", path = "crates/core", default-features = false }
-wasmi_ir = { version = "0.46.0", path = "crates/ir", default-features = false }
-wasmi_collections = { version = "0.46.0", path = "crates/collections", default-features = false }
-wasmi_c_api_impl = { version = "0.46.0", path = "crates/c_api" }
-wasmi_c_api_macros = { version = "0.46.0", path = "crates/c_api/macro" }
-wasmi_fuzz = { version = "0.46.0", path = "crates/fuzz" }
-wasmi_wast = { version = "0.46.0", path = "crates/wast" }
+wasmi = { version = "0.47.0", path = "crates/wasmi", default-features = false }
+wasmi_wasi = { version = "0.47.0", path = "crates/wasi", default-features = false }
+wasmi_core = { version = "0.47.0", path = "crates/core", default-features = false }
+wasmi_ir = { version = "0.47.0", path = "crates/ir", default-features = false }
+wasmi_collections = { version = "0.47.0", path = "crates/collections", default-features = false }
+wasmi_c_api_impl = { version = "0.47.0", path = "crates/c_api" }
+wasmi_c_api_macros = { version = "0.47.0", path = "crates/c_api/macro" }
+wasmi_fuzz = { version = "0.47.0", path = "crates/fuzz" }
+wasmi_wast = { version = "0.47.0", path = "crates/wast" }
 
 # wasm-tools dependencies
 wat = { version = "1.229.0", default-features = false }
